failles csrf reglee pour ne pas se faire hacker par lien

+++ views/admin/category/delete.php
@@ -7,6 +7,10 @@ Auth::check();
 $pdo = Connection::getPDO();
 $table = new CategoryTable($pdo);
 $category = $table->find($params['id']);
+if ($_SESSION['token'] !== $params['token']) {
+    header('Location :' . $router->url('admin_categories'));
+    exit();
+}
 (new CategoryAttachment)->detach($category);
 $table->delete($params['id']);
 header('Location: ' . $router->url('admin_categories') . '?delete=1');
