+++ views/admin/image/delete.php
@@ -7,6 +7,10 @@ Auth::check();
 $pdo = Connection::getPDO();
 $table = new ImageTable($pdo);
 $image = $table->find($params['id']);
+if ($_SESSION['token'] !== $params['token']) {
+    header('Location :' . $router->url('admin_images'));
+    exit();
+}
 $table->delete($params['id']); // Supprimer la photo de la bdd
 $link = UPLOAD_PATH . DIRECTORY_SEPARATOR . 'posts_multiple' . DIRECTORY_SEPARATOR . $image->getName();
 if (file_exists($link)) {
